Rename setupbase for tab completion
It is setupbase now
