+++ setup.py
@@ -14,7 +14,7 @@ DOCLINES = __doc__.split("\n")
 import sys
 from setuptools import setup, Extension
 sys.path.insert(0, '.')
-from setupbase import (find_packages, write_version_py, build_ext,
+from basesetup import (find_packages, write_version_py, build_ext,
                        StaticLibrary, CompilerDetection)
 
 try:
